fixed types

+++ src/genetic/algorithm.rs
@@ -74,10 +74,10 @@ where
         &self,
         population: &mut Vec<Box<TIndividual>>,
         population_size: usize,
-        children_count: u16,
+        children_count: u32,
     ) -> Result<Vec<Box<TIndividual>>, ()> {
         let mut rng = thread_rng();
-        let max_children_count = 1 + (rng.next_u32() as u16 % children_count);
+        let max_children_count = 1 + (rng.next_u32() % children_count);
         let mut mutants: Vec<_> = population
             .into_par_iter()
             .flat_map(|parent| {
